refactor: move letter spacing
